requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.0.1
altgraph==0.17.3
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
astroid==2.14.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
bardapi==0.1.33
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
blinker==1.6.2
blis==0.7.9
boto3==1.26.79
botocore==1.29.79
browser-cookie3==0.19.1
bs4==0.0.1
cachetools==5.3.0
catalogue==2.0.8
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
claude-api==1.0.17
cleanco==2.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
confection==0.0.3
configobj==5.0.8
contourpy==1.0.7
cryptography==39.0.1
cssselect==1.2.0
curl-cffi==0.5.9
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
dataclasses-json==0.5.14
dateparser==1.1.8
debugpy==1.6.5
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
demjson3==3.0.6
dictdiffer==0.9.0
dill==0.3.6
diskcache==5.4.0
distlib==0.3.6
distro==1.8.0
docopt==0.6.2
dpath==2.1.4
duckdb==0.8.1
duckdb-engine==0.6.8
dulwich==0.21.3
dvc==2.45.1
dvc-data==0.40.3
dvc-gdrive==2.19.1
dvc-http==2.30.2
dvc-objects==0.19.3
dvc-render==0.2.0
dvc-studio-client==0.5.1
dvc-task==0.1.11
dvclive==2.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
facebook-scraper==0.2.59
fake-useragent==1.1.1
Faker==17.6.0
fastcore==1.5.29
fastjsonschema==2.16.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
flatten-dict==0.4.2
flet==0.7.4
flet-core==0.7.4
flufl.lock==7.1.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
fst-pso==1.8.1
funcy==1.17
future==0.18.2
FuzzyTM==2.0.5
gensim==4.3.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.98.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
google-search-results==2.4.2
googleapis-common-protos==1.58.0
grandalf==0.8
greenlet==2.0.2
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
hugchat==0.3.1
hydra-core==1.3.2
hyperframe==6.0.1
ibis-framework==4.1.0
idna==3.4
imageio==2.28.0
importlib-metadata==6.0.0
ipykernel==6.19.4
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
iterative-telemetry==0.0.7
itsdangerous==2.1.2
jedi==0.18.2
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.2
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab==2.2.9
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
kombu==5.2.4
kora==0.9.20
lab==7.3
langchain==0.0.292
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.38
lazy-object-proxy==1.9.0
linkedin-scraper==2.11.0
lxml==4.9.2
lz4==4.3.2
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.7.3
matplotlib-inline==0.1.6
mccabe==0.7.0
metaflow==2.8.0
mglearn==0.1.9
miniful==0.0.6
mistune==2.0.4
multidict==6.0.4
multipledispatch==0.6.0
murmurhash==1.0.9
mypy-extensions==1.0.0
nanotime==0.5.2
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==3.0
newspaper3k==0.2.8
nlp==0.4.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.27.8
openpyxl==3.1.1
orjson==3.8.6
outcome==1.2.0
overpy==0.6
packaging==23.1
pandas==1.5.3
pandasai==1.2.2
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
parsy==2.1
pathspec==0.11.0
pathy==0.10.1
pefile==2023.2.7
perplexityai @ git+https://github.com/nathanrchn/perplexityai.git@fca9bb5220cc8d0f2277aa5b78a9ca0d83602fb8
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.13
platformdirs==2.6.2
plotly==5.15.0
preshed==3.0.8
prettytable==3.8.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
proto-plus==1.22.3
protobuf==4.22.0
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.4
pydeck==0.8.1b0
pydot==1.4.2
PyDrive2==1.15.1
pyee==8.2.2
pyFUME==0.2.25
pygit2==1.11.1
Pygments==2.14.0
pygtrie==2.5.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyLDAvis==3.3.1
pylint==2.16.2
Pympler==1.0.1
pyngrok==6.0.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.4
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.2
QtPy==2.3.1
regex==2022.10.31
repath==0.9.0
requests==2.28.2
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.10.0
scmrepo==0.1.11
scrapy-fake-useragent==1.4.4
seaborn==0.12.2
selenium==4.8.2
Send2Trash==1.8.0
sgmllib3k==1.0.0
shellingham==1.5.0.post1
shortuuid==1.0.11
shtab==1.5.8
simpful==2.9.0
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
SQLAlchemy==2.0.20
sqlalchemy-views==0.3.2
sqlglot==10.6.4
sqltrie==0.0.28
srsly==2.4.5
st-pages==0.4.3
stack-data==0.6.2
streamlit==1.27.2
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
textblob==0.17.1
thinc==8.1.6
threadpoolctl==3.1.0
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==3.4.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
trio==0.22.0
trio-websocket==0.9.2
txt2tags==3.8
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3.1
undetected-chromedriver==3.4.6
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
validators==0.20.0
vine==5.0.0
virtualenv==20.20.0
voluptuous==0.13.1
w3lib==2.1.1
wasabi==0.10.1
watchdog==2.3.1
wcwidth==0.2.5
webcolors==1.12
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.6.4
websockets==10.4
Werkzeug==2.2.3
wget==3.2
whylogs==1.1.27
whylogs-sketching==3.4.1.dev3
widgetsnbextension==4.0.7
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.2.0
yarg==0.1.9
yarl==1.8.2
zc.lockfile==3.0
zipp==3.15.0
